Backport: 2005-09-26  Sebastien Pouliot  <sebastien@ximian.com>
	* io.c: Apply patch from mono/mono#76192 (Can't write files past 2gb on AMD64 
	(x86_64)) for Brion on IRC (Dick approved it).

svn path=/branches/mono-1-1-9/mono/; revision=50809


Commit migrated from https://github.com/mono/mono/commit/bdf5931ff0deee9a8771603a0073f1da340cb2b5

+++ src/mono/mono/io-layer/io.c
@@ -491,7 +491,7 @@ static guint32 file_seek(gpointer handle, gint32 movedistance,
 			  offset, movedistance);
 #endif
 	} else {
-		offset=((gint64) *highmovedistance << 32) | (unsigned long)movedistance;
+		offset=((gint64) *highmovedistance << 32) | (guint32)movedistance;
 		
 #ifdef DEBUG
 		g_message("%s: setting offset to %lld 0x%llx (high %d 0x%x, low %d 0x%x)", __func__, offset, offset, *highmovedistance, *highmovedistance, movedistance, movedistance);
